Sredjivane funkcije: sitne ispravke u tekstovima zadataka, dodata neka resenja, sredjivani komentari

+++ resenja/1_KontrolaToka/1.4_Funkcije/praktikumi9/4_18.c
@@ -1,36 +1,29 @@
 #include <stdio.h>
 
-/* Funkcija racuna n-ti clan Fibonacijevog niza.
- * Clanovi ovog niza zadaju se rekurzivno tj. u zavisnosti od prethodnih clanova.
- * Fibonacijevi brojevi od 0. do 47. se mogu smestiti u tip int, a kako n moze uzimati vrednosti
- * od 1 do 50, povratni tip funkcije je long int. 
- */
-long int fibonaci(int n) {
+/* Funkcija racuna n-ti clan Fibonacijevog niza. */
+long unsigned fibonaci(int n) {
 
   int i;
 
-  /* f0 i f1 su prva dva clana niza */
-  int f0 = 1; 
-  int f1 = 1;
-  /* promenljiva u kojoj se cuvaju opsti clanovi: n+2, n+1. i n-ti clan */
+  /* Prva dva clana niza su jednaki 1. */
+  long unsigned f0 = 1; 
+  long unsigned f1 = 1;
+  /* Promenljive u kojima se cuvaju tekuci clanovi niza: n+2, n+1. i n-ti clan. */
   long int fn2, fn1, fn;
   
-  /* ukoliko treba vratiti nulti ili prvi clan,
-   * njih ne treba racunati
-   * jer su vec dati.
-   */
+  /* Nulti i prvi clan Fibonacijevog niza su poznati pa ih ne treba racunati u petlji. */
   if(n == 0 || n == 1)
     return 1;
 
-  /* postavljamo prethodne clanove niza */
+  /* Tekuci clanovi niza se postavljaju  na pocetne vrednosti. */
   fn = f0;
   fn1 = f1;
-  /* racunamo od drugog clana, pa dok ne dodjemo do n-tog */
+  /* Elementi niza se racunaju od drugog clana do n-tog. */
   for(i = 2; i <= n; i++) {
   
-    /* izracunamo n+2-i clan niza sabiranjem prethodna dva clana */
+    /* Naredni clan niza (element n+2-i) se dobija sabiranjem prethodna dva clana. */
     fn2 = fn1 + fn; 
-    /* promenimo prethodne clanove niza, zbog naredne iteracije */
+    /* Menjaju se vrednosti tekucih clanova niza zbog naredne iteracije. */
     fn =  fn1;
     fn1 = fn2;
   }
@@ -44,12 +37,12 @@ int main() {
   printf("Unesite broj n: ");
   scanf("%d", &n);
   
-  /* Provera vrednosti za broj n */
+  /* Provera vrednosti za broj n. */
   if(n <0 || n > 50) {
     printf("Greska: nedozvoljena vrednost!\n");
   }
   else{  
-  	printf("%ld\n", fibonaci(n));
+  	printf("%lu\n", fibonaci(n));
   }
   
   return 0;
